+++ lib/shift.rb
@@ -57,62 +57,4 @@ class Shift
     end
     key_counter
   end
-
-  def index_checker(index_check)
-    times = (index_check.abs / 27).to_i
-      index_check -= (27 * times)
-  end
-
-  def change_message(message, shift_hash)
-    key_counter = 0
-    shifted_message = (message.downcase.chars).map do |letter_or_space|
-    key_counter = key_counter_check(key_counter)
-      if ((letter_or_space.match(/^[[:alpha:][:blank:]]+$/)) == nil) || (
-        (shift_hash[key_counter].to_i) % 27 == 0)
-        letter_or_space
-      elsif ((('a'..'z').to_a << " ").index(letter_or_space) +(shift_hash[key_counter])) <= 26
-        alphabet_and_space_array[(alphabet_and_space_array.index(letter_or_space) +
-        shift_hash[key_counter])]
-      else
-        alphabet_and_space_array[index_checker(alphabet_and_space_array.index(
-        letter_or_space) + shift_hash[key_counter])]
-      end
-    end
-    shifted_message.join
-  end
-
-  def adjust_number(ciphertext_index, shift_number)
-    return shift_number if ciphertext_index == nil
-    if shift_number > ciphertext_index
-      times = (shift_number / 27).to_i
-      shift_number = shift_number - (27 * times)
-    end
-    shift_number
-  end
-
-  def unchange_message(message, shift_hash)
-    key_counter = 0
-    shifted_message = (message.downcase.chars).map do |letter_or_space|
-      key_counter = key_counter_check(key_counter)
-      decrypt_shift = adjust_number(alphabet_and_space_array.index(letter_or_space), shift_hash[key_counter])
-      #rf
-      if ((letter_or_space.match(/^[[:alpha:][:blank:]]+$/)) == nil) || (
-        (shift_hash[key_counter].to_i) % 27 == 0)
-        letter_or_space
-      elsif (alphabet_and_space_array.index(letter_or_space)) >= decrypt_shift
-          alphabet_and_space_array[(alphabet_and_space_array.index(letter_or_space) - decrypt_shift)]
-      else
-        alphabet_and_space_array[(alphabet_and_space_array.index(letter_or_space) - (decrypt_shift - 27)).abs]
-      end
-    end
-    shifted_message.join
-  end
-
-    def encrypt(message, key, date, shift_hash)
-      {encryption: change_message(message, shift_hash), key: key, date: date}
-    end
-
-    def decrypt(message, key, date, shift_hash)
-      {decryption: unchange_message(message, shift_hash), key: key, date: date}
-    end
 end
