+++ test/shift_test.rb
@@ -57,64 +57,4 @@ class ShiftTest < Minitest::Test
     assert_equal (4), @shift.key_counter_check(3)
     assert_equal (1), @shift.key_counter_check(4)
   end
-
-  def test_it_can_check_index_and_find_correct_place_in_alphabet_array
-    assert_equal (1), @shift.index_checker(55)
-    assert_equal (26), @shift.index_checker(53)
-    assert_equal (26), @shift.index_checker(26)
-  end
-
-  def test_it_can_change_a_message
-    assert_equal ("ihlruancu osh sscdlv."),
-    @shift.change_message("The eagle has landed.", {1=>16, 2=>27, 3=>34, 4=>45})
-    assert_equal ("qbj"), @shift.change_message("abc", {1=>16, 2=>27, 3=>34, 4=>45})
-    assert_equal ("q*j"), @shift.change_message("A*c", {1=>16, 2=>27, 3=>34, 4=>45})
-    assert_equal ("qahsrbitscjutdkvuelwvfmxwgnyxhozyip zjqa krbalsc"),
-    @shift.change_message("Aaaabbbbccccddddeeeeffffgggghhhhiiiijjjjkkkkllll",
-      {1=>16, 2=>27, 3=>34, 4=>45})
-  end
-
-  def test_it_can_adjust_shift_number_for_decypher
-    assert_equal (4), @shift.adjust_number(18, 4)
-    assert_equal (17), @shift.adjust_number(4, 71)
-  end
-
-  def test_it_can_unchange_a_message
-    assert_equal ("aaaabbbbccccddddeeeeffffgggghhhhiiiijjjjkkkkllll"),
-    @shift.unchange_message("qahsrbitscjutdkvuelwvfmxwgnyxhozyip zjqa krbalsc",
-      {1=>16, 2=>27, 3=>34, 4=>45})
-     assert_equal ("the eagle has landed."),
-    @shift.unchange_message("ihlruancu osh sscdlv.", {1=>16, 2=>27, 3=>34, 4=>45})
-    assert_equal ("abc"), @shift.unchange_message("qbj", {1=>16, 2=>27, 3=>34, 4=>45})
-      shift_hash = @shift.shift_numbers("02715", "040895")
-    assert_equal ("hello world"), @shift.unchange_message("keder ohulw", shift_hash)
-  end
-
-  def test_it_can_encrypt
-    assert_equal ({:encryption=>"fuhsmpsvpa !", :key=>"21234", :date=>"130120"}),
-    @shift.encrypt("Hello world!", "21234", "130120", {1=>25, 2=>16, 3=>23, 4=>34})
-    assert_equal ({:encryption=> "nocrgrlrdoa?", :key=>"02345", :date=>"130120"}),
-    @shift.encrypt("How are you?", "02345", "130120", {1=>06, 2=>27, 3=>34, 4=>45})
-    assert_equal ({:encryption=>"zugnsxpnpue?", :key=>"18341", :date=>"101020"}),
-    @shift.encrypt("How are you?", "18341", "101020", {1=>18, 2=>87, 3=>38, 4=>41})
-    key = "02715"
-    date = "040895"
-    shift_hash = @shift.shift_numbers(key, date)
-    assert_equal ({encryption: "keder ohulw", key: "02715", date: "040895"}),
-    @shift.encrypt("hello world", key, date, shift_hash)
-    key = "18341"
-    date = "101010"
-    shift_hash = @shift.shift_numbers(key, date)
-    assert_equal ({encryption: "zrcnsulnpra?", key: "18341", date: "101010"}),
-    @shift.encrypt("how are you?", key, date, shift_hash)
-  end
-
-  def test_it_can_decrypt
-    assert_equal ({:decryption=>"hello world!", :key=>"21234", :date=>"130120"}),
-    @shift.decrypt("fuhsmpsvpa !", "21234", "130120", {1=>25, 2=>16, 3=>23, 4=>34})
-    assert_equal ({:decryption=> "how are you?", :key=>"02345", :date=>"130120"}),
-    @shift.decrypt("nocrgrlrdoa?", "02345", "130120", {1=>06, 2=>27, 3=>34, 4=>45})
-    assert_equal ({:decryption=>"how are you?", :key=>"18341", :date=>"101020"}),
-    @shift.decrypt("zugnsxpnpue?", "18341", "101020", {1=>18, 2=>87, 3=>38, 4=>41})
-  end
 end
